pyproject.toml
[project]
name = "sophia-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas Pires",email = "lcpires7.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "supabase (>=2.23.2,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
